requirements.txt
wxPython==4.2.2
mysql-connector-python==8.0.28
